+++ src/Yaapii.Atoms/Enumerator/Joined.cs
@@ -20,61 +20,45 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
-using Yaapii.Atoms.Func;
-using Yaapii.Atoms.Scalar;
+using Yaapii.Atoms.List;
 
 #pragma warning disable NoProperties // No Properties
 #pragma warning disable CS1591
-
 namespace Yaapii.Atoms.Enumerator
 {
     /// <summary>
-    /// A <see cref="IEnumerator{T}"/> sorted by the given <see cref="Comparer{T}"/>.
+    /// Multiple <see cref="IEnumerator{T}"/> joined together.
     /// </summary>
-    /// <typeparam name="T">type of items in enumertor</typeparam>
-    public sealed class SortedEnumerator<T> : IEnumerator<T>
-        where T : IComparable<T>
+    /// <typeparam name="T">type of elements</typeparam>
+    public sealed class Joined<T> : IEnumerator<T>
     {
-        private readonly IScalar<IEnumerator<T>> _sorted;
+        private readonly IEnumerable<IEnumerator<T>> _list;
+        private readonly Queue<IEnumerator<T>> _buffer;
 
         /// <summary>
-        /// A <see cref="IEnumerator{T}"/> sorted by the given <see cref="Comparer{T}"/>.
+        /// Multiple <see cref="IEnumerator{T}"/> joined together.
         /// </summary>
-        /// <param name="cmp">comparer</param>
-        /// <param name="src">enumerator to sort</param>
-        public SortedEnumerator(Comparer<T> cmp, IEnumerator<T> src)
-        {
-            this._sorted =
-                new Sticky<IEnumerator<T>>(
-                    () =>
-                    {
-                        var items = new List<T>();
-                        while (src.MoveNext())
-                        {
-                            items.Add(src.Current);
-                        }
-                        items.Sort(cmp);
-
-                        return items.GetEnumerator();
-                    });
-        }
-
-        public void Dispose()
+        /// <param name="items">enumerables to join together</param>
+        public Joined(params IEnumerator<T>[] items) : this(new List<IEnumerator<T>>(items))
         { }
 
-        public Boolean MoveNext()
+        /// <summary>
+        /// Multiple <see cref="IEnumerator{T}"/> joined together.
+        /// </summary>
+        /// <param name="items">enumerables to join together</param>
+        public Joined(IEnumerable<IEnumerator<T>> items)
         {
-            return this._sorted.Value().MoveNext();
+            this._list = items;
+            this._buffer = new Queue<IEnumerator<T>>(items);
         }
 
         public T Current
         {
             get
             {
-                return this._sorted.Value().Current;
+                return this._buffer.Peek().Current;
             }
         }
 
@@ -82,13 +66,34 @@ namespace Yaapii.Atoms.Enumerator
         {
             get
             {
-                return this._sorted.Value().Current;
+                return Current;
             }
+        }
+
+        public void Dispose()
+        {
+
+        }
+
+        public bool MoveNext()
+        {
+            while(this._buffer.Count > 0 && !this._buffer.Peek().MoveNext())
+            {
+                this._buffer.Dequeue();
+            }
+
+            return this._buffer.Count > 0;
         }
 
         public void Reset()
         {
-            throw new NotSupportedException("#Reset() is not supported");
+            this._buffer.Clear();
+            var e = this._list.GetEnumerator();
+            while (e.MoveNext())
+            {
+                this._buffer.Enqueue(e.Current);
+            }
+            
         }
     }
 }
