+++ src/Yaapii.Atoms/Enumerator/Sibling.cs
@@ -36,7 +36,7 @@ namespace Yaapii.Atoms.Enumerator
     /// Element from position, starting with given item in a <see cref="IEnumerable{T}"/> fallback function <see cref="IFunc{In, Out}"/>.
     /// </summary>
     /// <typeparam name="T"></typeparam>
-    public sealed class SiblingEnumerator<T> : IScalar<T>
+    public sealed class Sibling<T> : IScalar<T>
         where T : IComparable<T>
     {
         /// <summary>
@@ -65,7 +65,7 @@ namespace Yaapii.Atoms.Enumerator
         /// <param name="src">the enumerator</param>
         /// <param name="item">the item to start with</param>
         /// <param name="fallback">the fallback to return if fails</param>
-        public SiblingEnumerator(IEnumerator<T> src, T item, T fallback) : this(
+        public Sibling(IEnumerator<T> src, T item, T fallback) : this(
             src,
             item,
             1,
@@ -79,7 +79,7 @@ namespace Yaapii.Atoms.Enumerator
         /// <param name="item">the item to start with</param>
         /// <param name="pos">position of the neighbour</param>
         /// <param name="fallback">the fallback to return if fails</param>
-        public SiblingEnumerator(IEnumerator<T> src, T item, int pos, T fallback) : this(
+        public Sibling(IEnumerator<T> src, T item, int pos, T fallback) : this(
             src,
             item,
             pos,
@@ -92,7 +92,7 @@ namespace Yaapii.Atoms.Enumerator
         /// <param name="src">source <see cref="IEnumerable{T}"/></param>
         /// <param name="pos">position</param>
         /// <param name="item">item to start with</param>
-        public SiblingEnumerator(IEnumerator<T> src, T item, int pos)
+        public Sibling(IEnumerator<T> src, T item, int pos)
         :
             this(
                 src,
@@ -118,7 +118,7 @@ namespace Yaapii.Atoms.Enumerator
         /// <param name="fbk">fallback function</param>
         /// <param name="pos">position</param>
         /// <param name="item">item to start with</param>
-        public SiblingEnumerator(
+        public Sibling(
             IEnumerator<T> src,
             T item,
             int pos,
