+++ tests/Yaapii.Atoms.Tests/Enumerator/LimitedTest.cs
@@ -24,45 +24,28 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using Xunit;
-using Yaapii.Atoms.List;
-using Yaapii.Atoms.Fail;
 using Yaapii.Atoms.Enumerable;
 using Yaapii.Atoms.Enumerator;
+using Yaapii.Atoms.List;
+using Yaapii.Atoms.Text;
 
 namespace Yaapii.Atoms.Enumerator.Tests
 {
-    public sealed class SkippedEnumeratorTest
+    public sealed class LimitedTest
     {
         [Fact]
-        public void SkipEnumerator()
+        public void LimitsContent()
         {
-            var skipped =
-                new List<string>(
-                    new EnumerableOf<string>(
-                        new SkippedEnumerator<string>(
-                            new EnumerableOf<string>(
-                                "one", "two", "three", "four"
-                            ).GetEnumerator(),
-                        2)));
-
-            Assert.True(new LengthOfEnumerator(skipped.GetEnumerator()).Value() == 2, "cannot skip elements");
-            Assert.False(skipped.Contains("one"), "cannot skip elements");
-            Assert.False(skipped.Contains("two"), "cannot skip elements");
-            Assert.True(skipped.Contains("three"), "cannot skip elements");
-            Assert.True(skipped.Contains("four"), "cannot skip elements");
-        }
-
-        [Fact]
-        public void CannotSkippedMoreThanExists()
-        {
-            Assert.False(
-                new SkippedEnumerator<string>(
-                    new EnumerableOf<string>(
-                        "one", "two"
-                    ).GetEnumerator(),
-                    2
-                ).MoveNext(),
-                "enumerates more elements than exist");
+            Assert.True(
+                new JoinedText(", ",
+                new EnumerableOf<IText>(
+                    new Mapped<int, IText>(
+                        new Limited<int>(
+                            new EnumerableOf<int>(1, 2, 3, 4).GetEnumerator(),
+                            2), 
+                        str => new TextOf(str + "")))).AsString() == "1, 2",
+            "cannot limit enumertor contents");
+                
         }
     }
 }
