+++ tests/Yaapii.Atoms.Tests/Enumerator/StickyTest.cs
@@ -23,27 +23,38 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading;
 using Xunit;
-using Yaapii.Atoms.List;
-using Yaapii.Atoms.Func;
-using Yaapii.Atoms.Enumerator;
+using Yaapii.Atoms;
 using Yaapii.Atoms.Enumerable;
+using Yaapii.Atoms.Enumerator;
+using Yaapii.Atoms.List;
+using Yaapii.Atoms.Scalar;
+using Yaapii.Atoms.Text;
 
 namespace Yaapii.Atoms.Enumerator.Tests
 {
-    public sealed class JoinedEnumeratorTest
+    public sealed class StickyTest
     {
         [Fact]
-        public void JoinsEnumerators()
+        public void IgnoresChangesInIterable()
         {
+            int count = 10;
             Assert.True(
-            new LengthOfEnumerator(
-                new JoinedEnumerator<IEnumerator<String>>(
-                    new MappedEnumerator<string, IEnumerator<string>>(
-                        new EnumerableOf<string>("x", "y", "z").GetEnumerator(),
-                        (input) => new EnumerableOf<string>(input).GetEnumerator()))
-                    ).Value() == 3,
-            "Can't concatenate mapped iterators together");
+                new JoinedText(
+                    ", ",
+                    new EnumerableOf<IText>(
+                        new Mapped<int, IText>(
+                            new Sticky<int>(
+                                new Limited<int>(
+                                    new Endless<int>(Interlocked.Increment(ref count)),
+                                    20
+                                )
+                            ),
+                            (number) => new TextOf(number + ""))))
+                            .AsString() == "11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11",
+                "cannot cache iterator");
+
         }
     }
 }
